Added base health. Fixed pathing. Changed tower placement
To place towers you now hold down control, alt, or shift and left click
to place towers of 3 different kinds.

The sprite for the base is now the same as before since collision was
wonky. Pathing now goes to the center of the sprite instead of to the
top left corner.

Base HP also implemented and creeps now die when they reach the base.
Currently they just subtract 5 lives each and there's nothing that
happens when the base dies (although that would be a trivial addition in
map.update())